// src/modules/finance/enums/account-receivable-bucket-situation.enum.ts
export enum AccountReceivableBucketSituationEnum {
  A_VENCER_0_30 = 'A_VENCER_0_30',
  A_VENCER_31_60 = 'A_VENCER_31_60',
  A_VENCER_61_90 = 'A_VENCER_61_90',
  A_VENCER_91_180 = 'A_VENCER_91_180',
  A_VENCER_181_360 = 'A_VENCER_181_360',
  A_VENCER_MAIOR_360 = 'A_VENCER_MAIOR_360',
  VENCIDOS_0_30 = 'VENCIDOS_0_30',
  VENCIDOS_31_60 = 'VENCIDOS_31_60',
  VENCIDOS_61_90 = 'VENCIDOS_61_90',
  VENCIDOS_91_180 = 'VENCIDOS_91_180',
  VENCIDOS_181_360 = 'VENCIDOS_181_360',
  VENCIDOS_MAIOR_360 = 'VENCIDOS_MAIOR_360',
}